+++ tests/test_component.py
@@ -19,7 +19,6 @@ from pytest_django.asserts import assertHTMLEqual, assertInHTML
 from django_components import (
     Component,
     ComponentRegistry,
-    ComponentView,
     Slot,
     SlotInput,
     all_components,
@@ -1240,12 +1239,12 @@ class TestComponentRender:
             def get_template_data(self, args, kwargs, slots, context):
                 return {"how": kwargs.pop("how")}
 
-            class View(ComponentView):
+            class View:
                 def get(self, request):
                     how = "via GET request"
 
-                    return self.component.render_to_response(
-                        context=RequestContext(self.request),
+                    return self.component_cls.render_to_response(  # type: ignore[attr-defined]
+                        context=RequestContext(request),
                         kwargs={"how": how},
                     )